2FA for encrypted session creation

+++ internal/adapters/encryption/session_builder/builder.go
@@ -2,9 +2,11 @@ package sessbldr
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 
 	domainErrors "go.openfort.xyz/shield/internal/core/domain/errors"
+	"go.openfort.xyz/shield/internal/core/domain/share"
 	"go.openfort.xyz/shield/internal/core/ports/builders"
 	"go.openfort.xyz/shield/internal/core/ports/repositories"
 	"go.openfort.xyz/shield/internal/core/ports/strategies"
@@ -16,18 +18,20 @@ type sessionBuilder struct {
 	encryptionPartsRepo    repositories.EncryptionPartsRepository
 	projectRepo            repositories.ProjectRepository
 	reconstructionStrategy strategies.ReconstructionStrategy
+	requireOTPCheck        bool
 }
 
-func NewEncryptionKeyBuilder(encryptionPartsRepo repositories.EncryptionPartsRepository, projectRepository repositories.ProjectRepository, reconstructionStrategy strategies.ReconstructionStrategy) builders.EncryptionKeyBuilder {
+func NewEncryptionKeyBuilder(encryptionPartsRepo repositories.EncryptionPartsRepository, projectRepository repositories.ProjectRepository, reconstructionStrategy strategies.ReconstructionStrategy, requireOTPCheck bool) builders.EncryptionKeyBuilder {
 	return &sessionBuilder{
 		encryptionPartsRepo:    encryptionPartsRepo,
 		projectRepo:            projectRepository,
 		reconstructionStrategy: reconstructionStrategy,
+		requireOTPCheck:        requireOTPCheck,
 	}
 }
 
 func (b *sessionBuilder) SetProjectPart(ctx context.Context, identifier string) error {
-	part, err := b.encryptionPartsRepo.Get(ctx, identifier)
+	data, err := b.encryptionPartsRepo.Get(ctx, identifier)
 	if err != nil {
 		if errors.Is(err, domainErrors.ErrEncryptionPartNotFound) {
 			return domainErrors.ErrInvalidEncryptionSession
@@ -35,12 +39,21 @@ func (b *sessionBuilder) SetProjectPart(ctx context.Context, identifier string) 
 		return err
 	}
 
+	var part share.EncryptionPart
+	if err := json.Unmarshal([]byte(data), &part); err != nil {
+		return err
+	}
+
+	if b.requireOTPCheck && !part.OTPVerified {
+		return domainErrors.ErrOTPVerificationRequired
+	}
+
 	err = b.encryptionPartsRepo.Delete(ctx, identifier)
 	if err != nil {
 		return err
 	}
 
-	b.projectPart = part
+	b.projectPart = part.EncPart
 	return nil
 }
